+++ engine/engine/src/main/groovy/com/virtusa/gto/nyql/engine/repo/QRepositoryImpl.groovy
@@ -5,18 +5,14 @@ import com.virtusa.gto.nyql.configs.Configurations
 import com.virtusa.gto.nyql.engine.exceptions.NyScriptExecutionException
 import com.virtusa.gto.nyql.engine.exceptions.NyScriptParseException
 import com.virtusa.gto.nyql.exceptions.NyException
-import com.virtusa.gto.nyql.model.QRepository
-import com.virtusa.gto.nyql.model.QScript
-import com.virtusa.gto.nyql.model.QScriptList
-import com.virtusa.gto.nyql.model.QScriptMapper
-import com.virtusa.gto.nyql.model.QScriptResult
-import com.virtusa.gto.nyql.model.QSession
-import com.virtusa.gto.nyql.model.QSource
+import com.virtusa.gto.nyql.model.*
+import groovy.transform.CompileStatic
 import org.codehaus.groovy.control.CompilationFailedException
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 import java.lang.reflect.Field
+import java.nio.file.Paths
 
 /**
  * @author IWEERARATHNA
@@ -50,7 +46,9 @@ class QRepositoryImpl implements QRepository {
         LOGGER.warn('All caches cleared in query repository!')
     }
 
-    QScript parse(String scriptId, QSession session) throws NyException {
+    @CompileStatic
+    QScript parse(String scriptIdGiven, QSession session) throws NyException {
+        String scriptId = resolveScriptId(scriptIdGiven, session.currentCallingFromScript())
         QSource src = mapper.map(scriptId)
 
         if (configurations.cacheGeneratedQueries() && caching.hasGeneratedQuery(scriptId)) {
@@ -69,12 +67,22 @@ class QRepositoryImpl implements QRepository {
             return script
 
         } catch (CompilationFailedException ex) {
-            throw new NyScriptParseException(scriptId, src.file, ex)
+            throw new NyScriptParseException(scriptId, src.id, ex)
         } catch (IOException ex) {
-            throw new NyScriptExecutionException(scriptId, src.file, ex)
+            throw new NyScriptExecutionException(scriptId, src.id, ex)
         }
     }
 
+    @CompileStatic
+    private static String resolveScriptId(String scriptId, String currentScript) {
+        if (scriptId.startsWith('./') || scriptId.startsWith('../')) {
+            Paths.get(currentScript).resolve('..').resolve(scriptId).normalize().toString()
+        } else {
+            scriptId
+        }
+    }
+
+    @CompileStatic
     protected void cacheIfSpecified(Script compiledScript, String scriptId, QScript script) {
         try {
             Field field = compiledScript.getClass().getDeclaredField(configurations.cachingIndicatorVarName())
@@ -89,12 +97,13 @@ class QRepositoryImpl implements QRepository {
         }
     }
 
+    @CompileStatic
     protected static QScript convertResult(String scriptId, Object res, QSession session) {
         if (res instanceof QResultProxy) {
             return new QScript(id: scriptId, proxy: (QResultProxy) res, qSession: session)
         } else if (res instanceof QScriptList) {
             res.id = scriptId
-            return res
+            return (QScript) res
         } else {
             new QScriptResult(id: scriptId, qSession: session, scriptResult: res)
         }