+++ tests/scripts/scripts/schema_script.groovy
@@ -3,8 +3,7 @@
  */
 $DSL.script {
 
-    RUN("scripts/ddls/temp_create")
-    RUN("scripts/ddls/temp_create")
+    RUN("./ddls/temp_create")
     RUN("scripts/ddls/temp_create")
 
     RUN("scripts/ddls/temp_drop")