+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## NyQL Changelog
 
+### v1.1.3
+ * Features & Enhancements
+   - Relative script paths [#13](https://github.com/VirtusaPolarisGTO/NyQL/issues/13)
+   - Make TARGET() table by default for starting tables in JOIN closures [#14](https://github.com/VirtusaPolarisGTO/NyQL/issues/14)
+   - Support sql ANY/SOME and ALL conditional operators [#12](https://github.com/VirtusaPolarisGTO/NyQL/issues/12)
+ 
+ * Bug fixes
+   - Cross joins does not work [#15](https://github.com/VirtusaPolarisGTO/NyQL/issues/15)
+   - INSERT queries with inline QUERY does not correctly identify parameters [#11](https://github.com/VirtusaPolarisGTO/NyQL/issues/11)
+   - INSERT queries, which specified as RETURN_KEYS, will never return keys if cached [#9](https://github.com/VirtusaPolarisGTO/NyQL/issues/9)
+   - Cannot use HAVING clause without GROUP BY clause [#16](https://github.com/VirtusaPolarisGTO/NyQL/issues/16)
+   
 ### v1.1.2
  * [Features and Bug fixes](https://github.com/VirtusaPolarisGTO/NyQL/issues?q=is%3Aissue+milestone%3Av1.1.2+is%3Aclosed)